[FE] REFACT: 코드를 CoinLogPage에서 CoinLog로 이동#1604
